Remove phrase "merge down, rebase up"
This phrase caused a lot of confusion in past deliveries. For it to be
useful, developers must visualize a visual diagram of their git
branches, drawn in a particular way. I think that's unlikely to be
helpful to many people.

We also have a lab that involves a git visualization that draw `master`
above `development`, directly contradicting this mnemonic.

Close #42. Close #41. Close #33.